Major improvements to type system in codegen

+++ EasyVision/src/main/kotlin/io/github/deltacv/easyvision/attribute/misc/ListAttribute.kt
@@ -1,5 +1,6 @@
 package io.github.deltacv.easyvision.attribute.misc
 
+import com.google.gson.JsonObject
 import imgui.ImGui
 import imgui.ImVec2
 import imgui.extension.imnodes.ImNodes
@@ -12,6 +13,11 @@ import io.github.deltacv.easyvision.attribute.TypedAttribute
 import io.github.deltacv.easyvision.codegen.CodeGen
 import io.github.deltacv.easyvision.codegen.GenValue
 import io.github.deltacv.easyvision.gui.style.rgbaColor
+import io.github.deltacv.easyvision.serialization.data.DataSerializable
+import io.github.deltacv.easyvision.serialization.data.adapter.dataSerializableToJsonObject
+import io.github.deltacv.easyvision.serialization.data.adapter.jsonObjectToDataSerializable
+import io.github.deltacv.easyvision.serialization.data.adapter.toJsonObject
+import io.github.deltacv.easyvision.serialization.ev.AttributeSerializationData
 
 open class ListAttribute(
     override val mode: AttributeMode,
@@ -22,7 +28,7 @@ open class ListAttribute(
     var sameLine: Boolean = false
 ) : TypedAttribute(Companion) {
 
-    companion object: Type {
+    companion object : Type {
         override val name = "List"
         override val allowsNew = false
 
@@ -32,13 +38,15 @@ open class ListAttribute(
 
     override var typeName = "[${elementType.name}]"
 
-    override val styleColor get() = if(elementType.isDefaultListColor) {
-        Companion.styleColor
-    } else elementType.listStyleColor
+    override val styleColor
+        get() = if (elementType.isDefaultListColor) {
+            Companion.styleColor
+        } else elementType.listStyleColor
 
-    override val styleHoveredColor get() = if(elementType.isDefaultListColor) {
-        Companion.styleHoveredColor
-    } else elementType.listStyleHoveredColor
+    override val styleHoveredColor
+        get() = if (elementType.isDefaultListColor) {
+            Companion.styleHoveredColor
+        } else elementType.listStyleHoveredColor
 
     val listAttributes = mutableListOf<TypedAttribute>()
     val deleteQueue = mutableListOf<TypedAttribute>()
@@ -54,40 +62,55 @@ open class ListAttribute(
 
     private val allowAod get() = allowAddOrDelete && fixedLength == null
 
+    private var serializationData: Data? = null
+
+    @Suppress("UNCHECKED_CAST")
     override fun onEnable() {
-        // oh god... (it's been only 10 minutes and i have already forgotten how this works)
-        if(previousLength != fixedLength) {
-            if(fixedLength != null && (previousLength == null || previousLength == 0)) {
-                repeat(fixedLength!!) {
-                    createElement()
-                }
-            } else if(previousLength != null || previousLength != 0) {
-                val delta = (fixedLength ?: 0) - (previousLength ?: 0)
+        if (serializationData != null) {
+            listAttributes.clear()
 
-                if(delta < 0) {
-                    repeat(-delta) {
-                        val last = listAttributes[listAttributes.size - 1]
-                        last.delete()
+            for (obj in serializationData!!.attributes) {
+                val elem = createElement(false)
+                jsonObjectToDataSerializable(obj, inst = elem as DataSerializable<Any>)
+                elem.enable()
+            }
 
-                        listAttributes.remove(last)
-                        deleteQueue.add(last)
+            serializationData = null
+        } else {
+            // oh god... (it's been only 10 minutes and i have already forgotten how this works)
+            if (previousLength != fixedLength) {
+                if (fixedLength != null && (previousLength == null || previousLength == 0)) {
+                    repeat(fixedLength!!) {
+                        createElement()
                     }
-                } else {
-                    repeat(delta) {
-                        if(deleteQueue.isNotEmpty()) {
-                            val last = deleteQueue[deleteQueue.size - 1]
-                            last.restore()
+                } else if (previousLength != null || previousLength != 0) {
+                    val delta = (fixedLength ?: 0) - (previousLength ?: 0)
 
-                            listAttributes.add(last)
-                            deleteQueue.remove(last)
-                        } else {
-                            createElement()
+                    if (delta < 0) {
+                        repeat(-delta) {
+                            val last = listAttributes[listAttributes.size - 1]
+                            last.delete()
+
+                            listAttributes.remove(last)
+                            deleteQueue.add(last)
+                        }
+                    } else {
+                        repeat(delta) {
+                            if (deleteQueue.isNotEmpty()) {
+                                val last = deleteQueue[deleteQueue.size - 1]
+                                last.restore()
+
+                                listAttributes.add(last)
+                                deleteQueue.remove(last)
+                            } else {
+                                createElement()
+                            }
                         }
                     }
-                }
-            } else {
-                for(attribute in listAttributes.toTypedArray()) {
-                    attribute.delete()
+                } else {
+                    for (attribute in listAttributes.toTypedArray()) {
+                        attribute.delete()
+                    }
                 }
             }
         }
@@ -98,9 +121,9 @@ open class ListAttribute(
     override fun draw() {
         super.draw()
 
-        for((i, attrib) in listAttributes.withIndex()) {
-            if(beforeHasLink != hasLink) {
-                if(hasLink) {
+        for ((i, attrib) in listAttributes.withIndex()) {
+            if (beforeHasLink != hasLink) {
+                if (hasLink) {
                     // delete attributes if a link has been created
                     attrib.delete()
                 } else {
@@ -110,11 +133,11 @@ open class ListAttribute(
                 }
             }
 
-            if(!hasLink) { // only draw attributes if there's not a link attached
+            if (!hasLink) { // only draw attributes if there's not a link attached
                 isDrawAttributeTextOverriden = true
                 drawAttributeText(i, attrib)
 
-                if(isDrawAttributeTextOverriden) {
+                if (isDrawAttributeTextOverriden) {
                     ImGui.sameLine()
                 } else {
                     attrib.inputSameLine = true
@@ -134,8 +157,8 @@ open class ListAttribute(
     }
 
     override fun value(current: CodeGen.Current): GenValue {
-        return if(mode == AttributeMode.INPUT) {
-            if(hasLink) {
+        return if (mode == AttributeMode.INPUT) {
+            if (hasLink) {
                 val linkedAttrib = linkedAttribute()
 
                 raiseAssert(
@@ -169,7 +192,7 @@ open class ListAttribute(
     override fun drawAttribute() {
         super.drawAttribute()
 
-        if(!hasLink && elementType.allowsNew && allowAod && mode == AttributeMode.INPUT) {
+        if (!hasLink && elementType.allowsNew && allowAod && mode == AttributeMode.INPUT) {
             // idk wat the frame height is, i just stole it from
             // https://github.com/ocornut/imgui/blob/7b8bc864e9af6c6c9a22125d65595d526ba674c5/imgui_widgets.cpp#L3439
 
@@ -179,16 +202,16 @@ open class ListAttribute(
 
             ImGui.sameLine(0.0f, style.itemInnerSpacingX * 2.0f)
 
-            if(ImGui.button("+", buttonSize, buttonSize)) { // creates a new element with the + button
+            if (ImGui.button("+", buttonSize, buttonSize)) { // creates a new element with the + button
                 // uses the "new" function from the attribute's companion Type
                 createElement()
             }
 
             // display the - button only if the attributes list is not empty
-            if(listAttributes.isNotEmpty()) {
+            if (listAttributes.isNotEmpty()) {
                 ImGui.sameLine(0.0f, style.itemInnerSpacingX)
 
-                if(ImGui.button("-", buttonSize, buttonSize)) {
+                if (ImGui.button("-", buttonSize, buttonSize)) {
                     // remove the last element from the list when - is pressed
                     listAttributes.removeLastOrNull()
                         ?.delete() // also delete it from the element id registry
@@ -199,20 +222,39 @@ open class ListAttribute(
 
     override fun acceptLink(other: Attribute) = other is ListAttribute && other.elementType == elementType
 
-    private fun createElement() {
+    private fun createElement(enable: Boolean = true): TypedAttribute {
         val count = listAttributes.size.toString()
-        val elementName = count + if(count.length == 1) " " else ""
+        val elementName = count + if (count.length == 1) " " else ""
 
         val element = elementType.new(AttributeMode.INPUT, elementName)
         element.parentNode = parentNode
-        element.enable() //enables the new element
+        if(enable) element.enable() //enables the new element
 
         element.drawType = false // hides the variable type
 
         listAttributes.add(element)
 
         onElementCreation(element)
+        return element
     }
 
-    open fun onElementCreation(element: Attribute) { }
+    open fun onElementCreation(element: Attribute) {}
+
+    override fun makeSerializationData(): AttributeSerializationData {
+        val objects = mutableListOf<JsonObject>()
+
+        for (attrib in listAttributes) {
+            objects.add(dataSerializableToJsonObject(attrib).asJsonObject)
+        }
+
+        return Data(objects)
+    }
+
+    override fun takeDeserializationData(data: AttributeSerializationData) {
+        if (data is Data)
+            serializationData = data
+    }
+
+    data class Data(var attributes: List<JsonObject>) : AttributeSerializationData()
+
 }
